# Conflicts:
#	backend/zupzup/src/docs/asciidoc/index.adoc
#	backend/zupzup/src/main/resources/templates/index.html

+++ backend/zupzup/src/main/java/com/twoez/zupzup/plogginglog/domain/Location.java
@@ -1,20 +1,21 @@
 package com.twoez.zupzup.plogginglog.domain;
 
 
+import java.math.BigDecimal;
 import lombok.Getter;
 
 @Getter
 public class Location {
 
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
-    public Location(Double latitude, Double longitude) {
+    public Location(BigDecimal latitude, BigDecimal longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
 
-    public static Location of(Double latitude, Double longitude) {
+    public static Location of(BigDecimal latitude, BigDecimal longitude) {
         return new Location(latitude, longitude);
     }
 }
